Modularize FBSDKBridgeAPI
Summary:
As title:
- Created FBSDKBridgeAPI class to handle all BridgeAPI related actions
- Move BridgeAPI related functions out of FBSDKApplicationDelegate
- Update Login and Share kits for above changes

Reviewed By: codytwinton

Differential Revision: D14378194

fbshipit-source-id: b96b96446ebf5d2377c91c995fe2f8ebfee9f307
