+++ FBSDKCoreKit/FBSDKCoreKit/Internal/FBSDKCoreKit+Internal.h
@@ -24,14 +24,10 @@
 #import "AppEvents/Codeless/FBSDKViewHierarchy.h"
 #import "AppEvents/Codeless/FBSDKCodelessMacros.h"
 #import "AppEvents/Codeless/FBSDKCodelessIndexer.h"
-#import "BridgeAPI/FBSDKBridgeAPIProtocol.h"
-#import "BridgeAPI/FBSDKBridgeAPIProtocolType.h"
-#import "BridgeAPI/FBSDKBridgeAPIRequest.h"
-#import "BridgeAPI/FBSDKBridgeAPIResponse.h"
-#import "BridgeAPI/FBSDKURLOpening.h"
 #import "Cryptography/FBSDKCrypto.h"
 #import "FBSDKAudioResourceLoader.h"
 #import "FBSDKContainerViewController.h"
+#import "BridgeAPI/FBSDKBridgeAPI.h"
 #import "FBSDKMonotonicTime.h"
 #import "FBSDKSystemAccountStoreAdapter.h"
 #import "FBSDKTriStateBOOL.h"
