+++ FBSDKShareKit/FBSDKShareKit/Internal/FBSDKLikeDialog.m
@@ -98,7 +98,7 @@
                                                                                           userInfo:nil];
     void (^networkCompletionBlock)(FBSDKBridgeAPIResponse *) = ^(FBSDKBridgeAPIResponse *response) {
       if (response.error.code == FBSDKErrorAppVersionUnsupported) {
-        [[FBSDKApplicationDelegate sharedInstance] openBridgeAPIRequest:webRequest
+        [[FBSDKBridgeAPI sharedInstance] openBridgeAPIRequest:webRequest
                                                 useSafariViewController:useSafariViewController
                                                      fromViewController:self.fromViewController
                                                         completionBlock:completionBlock];
@@ -106,12 +106,12 @@
         completionBlock(response);
       }
     };
-    [[FBSDKApplicationDelegate sharedInstance] openBridgeAPIRequest:nativeRequest
+    [[FBSDKBridgeAPI sharedInstance] openBridgeAPIRequest:nativeRequest
                                             useSafariViewController:useSafariViewController
                                                  fromViewController:self.fromViewController
                                                     completionBlock:networkCompletionBlock];
   } else {
-    [[FBSDKApplicationDelegate sharedInstance] openBridgeAPIRequest:webRequest
+    [[FBSDKBridgeAPI sharedInstance] openBridgeAPIRequest:webRequest
                                             useSafariViewController:useSafariViewController
                                                  fromViewController:self.fromViewController
                                                     completionBlock:completionBlock];
